requirements.txt
lxml
requests
boto3
botocore
pyyaml